server/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bidict==0.23.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.103.1
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.9.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.27.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.23.2
websockets==11.0.3
wsproto==1.2.0